figlet: add missing err != nil check
Before, the figlet tool did not validate arguments! This could lead to
some strange behavior. For instance, if you ran

    toolbag ... -figlet:template totally/fake/path ...

you'd get

    figlet: dial: unknown network

instead of an error message about the bogus template. Why? Because the
figlet tool's argument-parsing method returned an error before reaching
the code that intitalized the dialing information. Then, the
initialization method ignored the error returned by the argument-parsing
method, stumbling blindly ahead to Dial()ing the socket.

This change fixes the typo that caused the latter.